added support to git repository mirroring
users just need to create a .mirror file in the root of the repository
with the url of the remote repository and blogc-git-receiver with try
to `git push --mirror --force` to it.

please note that the unix user that runs blogc-git-receiver (usually
`blogc`) must be able to push to the requested repository. creating a
passwordless ssh key or adding an oauth token to the url may be
reasonable ways to do it.